requirements.txt
Flask==3.0.3
gunicorn==22.0.0
opencv-python-headless==4.10.0.84
Pillow==10.4.0
qrcode==7.4.2
python-dotenv==1.0.1